ui: enhance hub layout
