+++ src/app/hub-exchanges/pages/index.ts
@@ -1,1 +1,4 @@
 export * from './view-exchange/view-exchange.page';
+export * from './view-exchange-giftee/view-exchange-giftee.page';
+export * from './view-exchange-self/view-exchange-self.page';
+export * from './view-exchange-summary/view-exchange-summary.page';
