+++ src/app/hub-exchanges/pages/view-exchange-summary/view-exchange-summary.page.html
@@ -1,0 +1,147 @@
+<div *ngrxLet="exchange$ as exchange" class="row">
+    <div class="col-md-4">
+        <h3 class="mb-3 fw-normal text-body-secondary">Lo importante</h3>
+        <div class="card border-0 shadow-lg rounded-lg mb-4">
+            <div class="card-body">
+                <div class="list-group">
+                    <div
+                        class="list-group-item list-group-item-action border-0">
+                        <div class="d-flex gap-3">
+                            <i class="bi bi-calendar-date fs-1"></i>
+                            <div
+                                class="d-flex flex-column justify-content-center">
+                                <small>
+                                    El intercambio está programado para
+                                </small>
+                                <span class="fs-5">
+                                    {{exchange?.scheduledOn | date}}
+                                </span>
+                            </div>
+                        </div>
+                    </div>
+                    <div
+                        class="list-group-item list-group-item-action border-0">
+                        <div class="d-flex gap-3">
+                            <i class="bi bi-coin fs-1"></i>
+                            <div
+                                class="d-flex flex-column justify-content-center">
+                                <small>El presupuesto de regalos es de</small>
+                                <span class="fs-5">
+                                    {{exchange?.budget | currency:'MXN'}}
+                                </span>
+                            </div>
+                        </div>
+                    </div>
+                    <div
+                        *ngIf="exchange?.drawnOn"
+                        class="list-group-item list-group-item-action border-0">
+                        <div class="d-flex gap-3">
+                            <i class="bi bi-shuffle fs-1"></i>
+                            <div
+                                class="d-flex flex-column justify-content-center">
+                                <small>Este intercambio fue sorteado el</small>
+                                <span class="fs-5">
+                                    {{exchange?.drawnOn | date}}
+                                </span>
+                            </div>
+                        </div>
+                    </div>
+                    <div
+                        *ngIf="!exchange?.drawnOn"
+                        class="list-group-item list-group-item-action border-0">
+                        <div class="d-flex gap-3">
+                            <i class="bi bi-clock fs-1"></i>
+                            <div
+                                class="d-flex flex-column justify-content-center">
+                                <small
+                                    >Este intercambio no ha sido sorteado</small
+                                >
+                                <span class="fs-5"> Sorteo pendiente </span>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+        <h3 class="mb-3 fw-normal text-body-secondary">Participantes</h3>
+        <div class="card border-0 shadow-lg rounded-lg mb-4">
+            <div class="card-body">
+                <div class="list-group">
+                    <div
+                        *ngFor="let participant of participants$ | async"
+                        class="list-group-item list-group-item-action border-0">
+                        <div *ngIf="participant" class="d-flex gap-3">
+                            <i class="bi bi-person-circle fs-1"></i>
+                            <div
+                                class="d-flex flex-column justify-content-center flex-grow-1">
+                                <div
+                                    class="d-flex justify-content-between align-items-center">
+                                    <span class="fs-5">
+                                        {{participant?.user?.name}}
+                                    </span>
+                                    <small
+                                        *ngIf="participant?.acknowledgedOn"
+                                        class="mb-0 text-success">
+                                        CONFIRMADO
+                                    </small>
+                                    <small
+                                        *ngIf="!participant?.acknowledgedOn"
+                                        class="mb-0 text-warning">
+                                        PENDIENTE
+                                    </small>
+                                </div>
+                                <small>{{participant?.user?.email}}</small>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div *ngrxLet="self$ as self" class="col-md-8">
+        <ng-container *ngIf="exchange?.drawnOn">
+            <h3 class="mb-3 fw-normal text-body-secondary">
+                Conoce a tu destinatario!
+            </h3>
+            <div
+                class="card bg-success-subtle border-0 shadow-lg rounded-lg mb-4">
+                <div class="card-body text-center">
+                    <p class="fs-3 mb-3">El intercambio se ha sorteado!</p>
+                    <p class="text-body-secondary">
+                        Descubre quién es tu destinatario y mira su wish list
+                    </p>
+                    <a
+                        type="button"
+                        class="btn btn-outline-secondary btn-lg"
+                        [routerLink]="['/hub/exchanges', exchange?._id, 'giftee']">
+                        Quiero ver quién es!
+                    </a>
+                </div>
+            </div>
+        </ng-container>
+        <ng-container *ngIf="!self?.acknowledgedOn">
+            <h3 class="mb-3 fw-normal text-body-secondary">
+                Confirma tu participación!
+            </h3>
+            <div
+                class="card bg-warning-subtle border-0 shadow-lg rounded-lg mb-4">
+                <div class="card-body text-center">
+                    <p class="fs-3 mb-3">No has confirmado tu participación!</p>
+                    <p class="text-body-secondary">
+                        Hazlo ahora mismo para poder comezar tu wishlist
+                    </p>
+                    <button
+                        type="button"
+                        class="btn btn-outline-secondary btn-lg"
+                        (click)="acknowledge(exchange?._id ?? '')">
+                        Quiero participar!
+                    </button>
+                </div>
+            </div>
+        </ng-container>
+        <h3 class="mb-3 fw-normal text-body-secondary">Instrucciones</h3>
+        <div class="card border-0 shadow-lg rounded-lg mb-4">
+            <div class="card-body" [innerHTML]="exchange?.instructions"></div>
+        </div>
+    </div>
+</div>
